seperate restaurant and item classess from solution.py to different files
