+++ ProgrammingLanguages/Item.py
@@ -1,0 +1,7 @@
+class Item:
+	label = []
+	price = 0.0
+
+	def __init__(self, label, price):
+		self.label = label
+		self.price = price
